rename persistence dir
